Split and move wait strategies into own modules
