+++ src/wait-strategy/wait.ts
@@ -1,5 +1,7 @@
-import { HealthCheckWaitStrategy, Log, LogWaitStrategy, WaitStrategy } from "./wait-strategy";
+import { WaitStrategy } from "./wait-strategy";
 import { HttpWaitStrategy } from "./http-wait-strategy";
+import { HealthCheckWaitStrategy } from "./health-check-wait-strategy";
+import { Log, LogWaitStrategy } from "./log-wait-strategy";
 
 export class Wait {
   public static forLogMessage(message: Log | RegExp, times = 1): WaitStrategy {
